@@ -13,48 +13,52 @@
     "upgrade-interactive": "npm-check --update"
   },
   "dependencies": {
-    "@material-ui/core": "4.11.3",
+    "@emotion/react": "11.4.1",
+    "@emotion/styled": "11.3.0",
+    "@material-ui/core": "4.12.3",
     "@material-ui/icons": "4.11.2",
-    "classnames": "^2.2.6",
-    "d3-array": "2.12.0",
-    "d3-axis": "2.1.0",
-    "d3-scale": "3.2.3",
-    "d3-selection": "2.0.0",
-    "d3-shape": "2.1.0",
-    "d3-time": "2.0.0",
-    "d3-transition": "2.0.0",
-    "firebase": "8.3.0",
+    "@mui/icons-material": "5.0.4",
+    "@mui/material": "5.0.4",
+    "@mui/styles": "5.0.1",
+    "classnames": "2.3.1",
+    "d3-array": "3.1.1",
+    "d3-axis": "3.0.0",
+    "d3-scale": "4.0.2",
+    "d3-selection": "3.0.0",
+    "d3-shape": "3.0.1",
+    "d3-time": "3.0.0",
+    "d3-transition": "3.0.1",
+    "firebase": "9.1.3",
     "moment": "2.29.1",
     "prop-types": "^15.7.2",
-    "react": "17.0.1",
-    "react-dom": "17.0.1",
-    "react-firebaseui": "5.0.0-canary1",
-    "react-redux": "7.2.2",
-    "react-router-dom": "5.2.0",
-    "redux": "4.0.5"
+    "react": "17.0.2",
+    "react-dom": "17.0.2",
+    "react-redux": "7.2.5",
+    "react-router-dom": "5.3.0",
+    "redux": "4.1.1"
   },
   "devDependencies": {
-    "@babel/core": "7.13.10",
-    "@babel/plugin-transform-runtime": "7.13.10",
-    "@babel/preset-env": "7.13.10",
-    "@babel/preset-react": "7.12.13",
+    "@babel/core": "7.15.8",
+    "@babel/plugin-transform-runtime": "7.15.8",
+    "@babel/preset-env": "7.15.8",
+    "@babel/preset-react": "7.14.5",
     "babel-loader": "8.2.2",
-    "clean-webpack-plugin": "3.0.0",
-    "eslint": "7.22.0",
+    "clean-webpack-plugin": "4.0.0",
+    "eslint": "7.32.0",
     "eslint-config-airbnb": "18.2.1",
     "eslint-loader": "4.0.2",
-    "eslint-plugin-import": "2.22.1",
+    "eslint-plugin-import": "2.25.2",
     "eslint-plugin-jsx-a11y": "6.4.1",
-    "eslint-plugin-react": "7.22.0",
+    "eslint-plugin-react": "7.26.1",
     "eslint-plugin-react-hooks": "4.2.0",
-    "firebase-tools": "9.6.1",
-    "html-webpack-plugin": "5.3.1",
-    "jest": "26.6.3",
+    "firebase-tools": "9.20.0",
+    "html-webpack-plugin": "5.4.0",
+    "jest": "27.2.5",
     "npm-check": "5.9.2",
-    "webpack": "5.26.1",
-    "webpack-cli": "4.5.0",
-    "webpack-dev-server": "3.11.2",
-    "webpack-merge": "5.7.3",
-    "webpack-notifier": "1.13.0"
+    "webpack": "5.58.2",
+    "webpack-cli": "4.9.0",
+    "webpack-dev-server": "4.3.1",
+    "webpack-merge": "5.8.0",
+    "webpack-notifier": "1.14.1"
   }
 }
