@@ -1,15 +1,15 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import Grid from '@material-ui/core/Grid';
-import Radio from '@material-ui/core/Radio';
-import AddIcon from '@material-ui/icons/Add';
-import RemoveIcon from '@material-ui/icons/Remove';
+import Grid from '@mui/material/Grid';
+import Radio from '@mui/material/Radio';
+import AddIcon from '@mui/icons-material/Add';
+import RemoveIcon from '@mui/icons-material/Remove';
 
 function BalanceRadio({ isPositive, onChange }) {
   return (
-    <Grid container direction="row" justify="space-around" alignItems="center" spacing={2}>
+    <Grid container direction="row" justifyContent="space-around" alignItems="center" spacing={2}>
       <Grid item>
-        <Grid container direction="row" justify="space-around" alignItems="center">
+        <Grid container direction="row" justifyContent="space-around" alignItems="center">
           <Grid item>
             <Radio
               checked={isPositive}
@@ -24,7 +24,7 @@ function BalanceRadio({ isPositive, onChange }) {
         </Grid>
       </Grid>
       <Grid item>
-        <Grid container direction="row" justify="space-around" alignItems="center">
+        <Grid container direction="row" justifyContent="space-around" alignItems="center">
           <Grid item>
             <Radio
               checked={!isPositive}
