@@ -1,13 +1,11 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import {
-  Grid,
-  CircularProgress,
-} from '@material-ui/core';
+import CircularProgress from '@mui/material/CircularProgress';
+import Grid from '@mui/material/Grid';
 
 function ProgressWait({ marginTop }) {
   return (
-    <Grid container direction="row" align="center" justify="center" style={{ marginTop }}>
+    <Grid container direction="row" align="center" justifyContent="center" style={{ marginTop }}>
       <Grid item>
         <CircularProgress />
       </Grid>
