@@ -1,12 +1,22 @@
-import firebase from './index';
+import {
+  getFirestore,
+  query,
+  where,
+  getDoc,
+  addDoc,
+  setDoc,
+  updateDoc,
+  deleteDoc,
+  onSnapshot,
+} from 'firebase/firestore';
 
+import firebaseApp from './index';
 import { getCurrentUser } from './auth';
-
 import { recurrentRef, punctualRef, settingRef } from './ref';
 
-const DB = firebase.firestore();
+const db = getFirestore(firebaseApp);
 
-function saveDocumentKey(doc, key) {
+function saveDocumentKey(doc, key = 'id') {
   return ({ ...doc.data(), [key]: doc.id });
 }
 
@@ -14,68 +24,76 @@ function snapshotToArray(querySnapshot, key) {
   return querySnapshot.docs.map(doc => saveDocumentKey(doc, key));
 }
 
+async function fetchDoc(docRef) {
+  return saveDocumentKey(await getDoc(docRef));
+}
+
 //
 // Listen
 //
 function listenAllRecurrentEntries(callback, uid) {
-  return recurrentRef(DB).where('authorId', '==', uid)
-    .onSnapshot(querySnapshot => callback(snapshotToArray(querySnapshot, 'id')));
+  return onSnapshot(
+    query(recurrentRef(db), where('authorId', '==', uid)),
+    querySnapshot => callback(snapshotToArray(querySnapshot)),
+  );
 }
 
 function listenPunctualEntries(callback, startDate, endDate, uid) {
-  return punctualRef(DB)
-    .where('date', '>=', startDate)
-    .where('date', '<=', endDate)
-    .where('authorId', '==', uid)
-    .onSnapshot(querySnapshot => callback(snapshotToArray(querySnapshot, 'id')));
+  return onSnapshot(
+    query(punctualRef(db), where('date', '>=', startDate), where('date', '<=', endDate), where('authorId', '==', uid)),
+    querySnapshot => callback(snapshotToArray(querySnapshot)),
+  );
 }
 
 function listenSettings(callback) {
-  return settingRef(DB, getCurrentUser().uid).onSnapshot(doc => callback(doc.data()));
+  return onSnapshot(
+    settingRef(db, getCurrentUser().uid),
+    doc => callback(saveDocumentKey(doc)),
+  );
 }
 
 //
 // Fetch
 //
 function fetchSettings() {
-  return settingRef(DB, getCurrentUser().uid).get().then(doc => doc.data());
+  return fetchDoc(settingRef(db, getCurrentUser().uid));
 }
 
 //
 // Push
 //
 function pushPunctualEntry(entry) {
-  return punctualRef(DB).add(entry);
+  return addDoc(punctualRef(db), entry);
 }
 
 function pushRecurrentEntry(entry) {
-  return recurrentRef(DB).add(entry);
+  return addDoc(recurrentRef(db), entry);
 }
 
 function pushSettingsCurrency(currency) {
-  return settingRef(DB, getCurrentUser().uid).set({ currency }, { merge: true });
+  return setDoc(settingRef(db, getCurrentUser().uid), { currency }, { merge: true });
 }
 
 //
 // Update
 //
 function updatePunctualEntry(entryId, entry) {
-  return punctualRef(DB, entryId).update(entry);
+  return updateDoc(punctualRef(db, entryId), entry);
 }
 
 function updateRecurrentEntry(entryId, entry) {
-  return recurrentRef(DB, entryId).update(entry);
+  return updateDoc(recurrentRef(db, entryId), entry);
 }
 
 //
 // Delete
 //
 function deletePunctualEntry(entryId) {
-  return punctualRef(DB, entryId).delete();
+  return deleteDoc(punctualRef(db, entryId));
 }
 
 function deleteRecurrentEntry(entryId) {
-  return recurrentRef(DB, entryId).delete();
+  return deleteDoc(recurrentRef(db, entryId));
 }
 
 export {
