@@ -1,12 +1,9 @@
-import firebase from 'firebase/app';
-
-import 'firebase/auth';
-import 'firebase/firestore';
-import 'firebase/performance';
+import { initializeApp } from 'firebase/app';
+import { getPerformance } from 'firebase/performance';
 
 const config = require('./config.json');
 
-firebase.initializeApp(config);
-firebase.performance();
+const firebaseApp = initializeApp(config);
+getPerformance(firebaseApp);
 
-export default firebase;
+export default firebaseApp;
